// Backend/src/db/types/public/Cidades.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { EstadosCodEstado } from './Estados';

/** Identifier type for public.Cidades */
export type CidadesCodCidade = number & { __brand: 'public.Cidades' };

/** Represents the table public.Cidades */
export default interface Cidades {
  codCidade: CidadesCodCidade;

  codEstado: EstadosCodEstado;

  cidade: string;

  ddd: string | null;

  dtCriacao: Date;

  dtEdicao: Date;

  codUser: number | null;

  isActive: boolean;
}

/** Represents the initializer for the table public.Cidades */
export interface CidadesInitializer {
  codCidade?: CidadesCodCidade;

  codEstado: EstadosCodEstado;

  cidade: string;

  ddd?: string | null;

  /** Default value: CURRENT_TIMESTAMP */
  dtCriacao?: Date;

  /** Default value: CURRENT_TIMESTAMP */
  dtEdicao?: Date;

  codUser?: number | null;

  /** Default value: true */
  isActive?: boolean;
}

/** Represents the mutator for the table public.Cidades */
export interface CidadesMutator {
  codCidade?: CidadesCodCidade;

  codEstado?: EstadosCodEstado;

  cidade?: string;

  ddd?: string | null;

  dtCriacao?: Date;

  dtEdicao?: Date;

  codUser?: number | null;

  isActive?: boolean;
}